引入SQL连接池,测试成功

+++ src/config.rs
@@ -22,7 +22,7 @@ pub struct Config{
     pub opt_linger:u32,
 
     //数据库连接池数量
-    pub sql_num:u32,
+    pub sql_num:usize,
 
     //线程池内的线程数量
     pub thread_num:u32,
@@ -44,7 +44,7 @@ impl Config{
         //传入是否优雅关闭连接
         self.opt_linger = value_t!(args,"opt_linger",u32).unwrap_or(0);
         //传入数据库连接数量
-        self.sql_num = value_t!(args,"sql_num",u32).unwrap_or(8);
+        self.sql_num = value_t!(args,"sql_num",usize).unwrap_or(8);
         //传入线程数量
         self.thread_num = value_t!(args,"thread_num",u32).unwrap_or(8);
         //传入是否关闭日志
